requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
drf-spectacular==0.27.1
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0
pytest==8.0.0
pytest-django==4.8.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0
isort==5.13.2 